refactor(useMessageHandling): split logic into hooks
